refactor: split ingest/service.py into per-adapter files (#188)
The monolithic service.py (~1250 lines) contained 4 adapter classes,
utility functions, and the orchestrator. This violates single-responsibility
and makes the ingest subsystem hard to navigate.

New structure:
- ingest/utils.py: shared utilities (compute_hash, chunk_text, etc.)
- ingest/adapters/url.py: URLAdapter
- ingest/adapters/pdf.py: PDFAdapter + docling/fitz/vision helpers
- ingest/adapters/text.py: TextAdapter
- ingest/adapters/youtube.py: YouTubeAdapter
- ingest/adapters/__init__.py: re-exports for convenience
- ingest/service.py: IngestService orchestrator + backward-compat re-exports

All existing imports continue to work via re-exports in service.py.
Test monkeypatching updated to target the actual adapter modules.

[skip-doc-check]
